backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-rest-auth[with-social]>=5.1.0",
    "django-environ>=0.11.2",
    "django-leaflet>=0.31.0",
    "django-rest-framework>=0.1.0",
    "django>=5.0.0",
    "drf-extra-fields>=3.7.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "six>=1.17.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "psycopg2-binary>=2.9.9",
    "daphne>=4.0.0",
    "djangorestframework-gis>=1.0.0",
]
